internal/ethapi: avoid int overflow in GetTransactionReceipt (#26911)

+++ internal/ethapi/api.go
@@ -2231,7 +2231,7 @@ func (s *PublicTransactionPoolAPI) GetTransactionReceipt(ctx context.Context, ha
 	if err != nil {
 		return nil, err
 	}
-	if len(receipts) <= int(index) {
+	if uint64(len(receipts)) <= index {
 		return nil, nil
 	}
 	receipt := receipts[index]
